move main/test.h and util/TestUtils.h files to test directory
