+++ src/bucket/BucketTests.cpp
@@ -18,8 +18,8 @@
 #include "herder/LedgerCloseData.h"
 #include "lib/catch.hpp"
 #include "main/Application.h"
-#include "main/test.h"
 #include "ledger/LedgerTestUtils.h"
+#include "test/test.h"
 #include "util/Fs.h"
 #include "util/Logging.h"
 #include "util/Timer.h"
