+++ src/database/DatabaseTests.cpp
@@ -6,8 +6,8 @@
 #include "database/Database.h"
 #include "main/Application.h"
 #include "main/Config.h"
-#include "main/test.h"
 #include "crypto/Hex.h"
+#include "test/test.h"
 #include "util/Logging.h"
 #include "util/Timer.h"
 #include "util/TmpDir.h"
