+++ src/herder/HerderTests.cpp
@@ -8,8 +8,8 @@
 #include "main/Config.h"
 #include "simulation/Simulation.h"
 #include "test/TestAccount.h"
+#include "test/test.h"
 
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "crypto/SHA.h"
 #include "test/TxTests.h"
