+++ src/history/HistoryTests.cpp
@@ -6,7 +6,6 @@
 #include "history/HistoryManager.h"
 #include "history/HistoryArchive.h"
 #include "history/HistoryWork.h"
-#include "main/test.h"
 #include "main/ExternalQueue.h"
 #include "main/Config.h"
 #include "main/PersistentState.h"
@@ -18,6 +17,7 @@
 #include "util/Logging.h"
 #include "util/Timer.h"
 #include "util/TmpDir.h"
+#include "test/test.h"
 #include "test/TxTests.h"
 #include "ledger/LedgerManager.h"
 #include "process/ProcessManager.h"
