+++ src/ledger/LedgerDeltaTests.cpp
@@ -3,13 +3,13 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "util/Timer.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "ledger/LedgerDelta.h"
 #include "main/Application.h"
 #include "LedgerTestUtils.h"
 #include "ledger/LedgerManager.h"
 #include "ledger/AccountFrame.h"
+#include "test/test.h"
 
 using namespace stellar;
