+++ src/ledger/LedgerHeaderTests.cpp
@@ -5,11 +5,11 @@
 #include "main/Application.h"
 #include "util/Timer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
 #include "ledger/LedgerManager.h"
 #include "herder/LedgerCloseData.h"
+#include "test/test.h"
 #include "xdrpp/marshal.h"
 
 #include "main/Config.h"
