+++ src/ledger/LedgerPerformanceTests.cpp
@@ -5,7 +5,6 @@
 #include "main/Application.h"
 #include "util/Timer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
 #include "ledger/LedgerManager.h"
@@ -17,6 +16,7 @@
 #include "simulation/Simulation.h"
 #include "herder/LedgerCloseData.h"
 #include "bucket/BucketManager.h"
+#include "test/test.h"
 #include "util/optional.h"
 #include "util/Math.h"
 #include "util/SociNoWarnings.h"
