+++ src/overlay/FloodTests.cpp
@@ -6,7 +6,6 @@
 #include "TCPPeer.h"
 #include "lib/catch.hpp"
 #include "main/Application.h"
-#include "main/test.h"
 #include "overlay/PeerDoor.h"
 #include "main/Config.h"
 #include "util/Logging.h"
@@ -14,6 +13,7 @@
 #include "overlay/OverlayManager.h"
 #include "simulation/Topologies.h"
 #include "test/TestAccount.h"
+#include "test/test.h"
 #include "test/TxTests.h"
 #include "herder/Herder.h"
 #include "ledger/LedgerDelta.h"
