+++ src/overlay/OverlayTests.cpp
@@ -5,7 +5,6 @@
 #include "main/Application.h"
 #include "overlay/LoopbackPeer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
 #include "util/Timer.h"
@@ -15,6 +14,7 @@
 #include "overlay/PeerRecord.h"
 #include "overlay/OverlayManagerImpl.h"
 #include "overlay/TCPPeer.h"
+#include "test/test.h"
 #include "BanManager.h"
 
 #include "medida/metrics_registry.h"
