+++ src/overlay/TCPPeerTests.cpp
@@ -6,9 +6,9 @@
 #include "TCPPeer.h"
 #include "lib/catch.hpp"
 #include "main/Application.h"
-#include "main/test.h"
 #include "overlay/PeerDoor.h"
 #include "main/Config.h"
+#include "test/test.h"
 #include "util/Logging.h"
 #include "simulation/Simulation.h"
 #include "overlay/OverlayManager.h"
