+++ src/process/ProcessTests.cpp
@@ -6,8 +6,8 @@
 #include "main/Config.h"
 #include "main/Application.h"
 #include "xdrpp/autocheck.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
+#include "test/test.h"
 #include "util/Logging.h"
 #include "util/Timer.h"
 #include "lib/util/format.h"
