+++ src/simulation/Simulation.cpp
@@ -7,9 +7,9 @@
 #include "herder/Herder.h"
 #include "ledger/LedgerManager.h"
 #include "main/Application.h"
-#include "main/test.h"
 #include "overlay/OverlayManager.h"
 #include "overlay/PeerRecord.h"
+#include "test/test.h"
 #include "util/Logging.h"
 #include "util/Math.h"
 #include "util/types.h"
