+++ src/transactions/ChangeTrustTests.cpp
@@ -5,13 +5,13 @@
 #include "main/Application.h"
 #include "overlay/LoopbackPeer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
-#include "util/TestUtils.h"
 #include "lib/json/json.h"
 #include "test/TestAccount.h"
 #include "test/TestExceptions.h"
+#include "test/TestUtils.h"
+#include "test/test.h"
 #include "test/TxTests.h"
 
 using namespace stellar;
