+++ src/transactions/OfferTests.cpp
@@ -6,15 +6,15 @@
 #include "main/Config.h"
 #include "overlay/LoopbackPeer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "lib/util/uint128_t.h"
 #include "util/Logging.h"
 #include "test/TestAccount.h"
 #include "test/TestExceptions.h"
+#include "test/TestUtils.h"
+#include "test/test.h"
 #include "test/TxTests.h"
 #include "transactions/OfferExchange.h"
-#include "util/TestUtils.h"
 #include "util/Timer.h"
 #include "database/Database.h"
