+++ src/transactions/SetOptionsTests.cpp
@@ -8,13 +8,13 @@
 #include "util/Timer.h"
 #include "overlay/LoopbackPeer.h"
 #include "util/make_unique.h"
-#include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
 #include "test/TestAccount.h"
 #include "test/TestExceptions.h"
+#include "test/TestUtils.h"
+#include "test/test.h"
 #include "test/TxTests.h"
-#include "util/TestUtils.h"
 #include "transactions/TransactionFrame.h"
 #include "ledger/LedgerDelta.h"
